server/ner/requirements.txt
absl-py
astunparse
cachetools
certifi
charset-normalizer
filelock
flatbuffers
fsspec
gast
google-auth
google-auth-oauthlib
google-pasta
grpcio
h5py
huggingface-hub
idna
importlib-metadata
Jinja2
joblib
keras
libclang
Markdown
MarkupSafe
mpmath
networkx
numpy
oauthlib
opt-einsum
packaging
pandas
Pillow
protobuf
pyasn1
pyasn1-modules
python-dateutil
pytz
pymongo
PyYAML
regex
requests
requests-oauthlib
rsa
safetensors
scikit-learn
scipy
six
sympy
tensorboard
tensorboard-data-server
tensorflow-addons
tensorflow-cpu
tensorflow-estimator
tensorflow-io-gcs-filesystem
termcolor
threadpoolctl
tokenizers
torch
torchaudio
torchvision
tqdm
transformers
typeguard
typing_extensions
tzdata
urllib3
Werkzeug
wrapt
zipp
